Refector async test utils and rename files
