+++ tests/async_utils.py
@@ -1,30 +1,45 @@
 """
-Drop-in replacement for absence of AsyncMock in python < 3.8.
-
-Based on https://github.com/timsavage/asyncmock
+ijtiset
 """
-
+import asyncio
 from unittest.mock import NonCallableMock, CallableMixin
 
+import pytest
+import pytest_asyncio
+try:
+    asyncio_fixture_decorator = pytest_asyncio.fixture
+except AttributeError:
+    asyncio_fixture_decorator = pytest.fixture
+
+class AwaitableTestFixture(object):
+    """Class implementing an awaitable test fixture."""
+    def __init__(self, awaitable_cls=None):
+        self.awaitable_cls = awaitable_cls
+
+    def __await__(self):
+
+        async def closure():
+            awaitables = [attr for attr in self.__dict__.values() if isinstance(
+                attr,  self.awaitable_cls
+            )]
+            await asyncio.gather(*awaitables)
+            return self
+        
+        return closure().__await__() 
 
 class AsyncCallableMixin(CallableMixin):
 
-    def __init__(_mock_self, not_async=False, *args, **kwargs):
+    def __init__(_mock_self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        _mock_self.not_async = not_async
         _mock_self.aenter_return_value = _mock_self
         _mock_self._await_count = 0
 
     def __call__(_mock_self, *args, **kwargs):
-        if _mock_self.not_async:
+        async def wrapper():
+            _mock_self._await_count += 1
             _mock_self._mock_check_sig(*args, **kwargs)
             return _mock_self._mock_call(*args, **kwargs)
-        else:
-            async def wrapper():
-                _mock_self._await_count += 1
-                _mock_self._mock_check_sig(*args, **kwargs)
-                return _mock_self._mock_call(*args, **kwargs)
-            return wrapper()
+        return wrapper()
 
     async def __aenter__(_mock_self):
         return _mock_self.aenter_return_value
@@ -51,12 +66,7 @@ class AsyncCallableMixin(CallableMixin):
 
 class AsyncMock(AsyncCallableMixin, NonCallableMock):
     """
-    Create a new `AsyncMock` object. `AsyncMock` several options that extends
-    the behaviour of the basic `Mock` object:
-    * `not_async`: This is a boolean flag used to indicate that when the mock
-      is called it should not return a normal Mock instance to make the mock
-      non-awaitable. If this flag is set the mock reverts to the default
-      behaviour of a `Mock` instance.
-    All other arguments are passed directly through to the underlying `Mock`
-    object.
+    Drop-in replacement for absence of AsyncMock in python < 3.8.
+
+    Based on https://github.com/timsavage/asyncmock
     """
